+++ src/main/java/uk/q3c/krail/option/api/OptionCacheConfig.java
@@ -14,8 +14,6 @@
 package uk.q3c.krail.option.api;
 
 import com.google.inject.BindingAnnotation;
-import uk.q3c.krail.core.eventbus.SessionBus;
-import uk.q3c.krail.core.guice.vsscope.VaadinSessionScope;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -24,9 +22,8 @@ import static java.lang.annotation.ElementType.*;
 import static java.lang.annotation.RetentionPolicy.*;
 
 /**
- * Binding annotation to identify an Event Bus as being global to the application (to distinguish from event buses which have {@link VaadinSessionScope} -
- * the latter are annotated with {@link SessionBus}
- * <p>
+ * Guice binding annotation to identify the Guava cache configuration to be used by Option
+ *
  * Created by David Sowerby on 06/02/15.
  */
 @BindingAnnotation