+++ src/main/java/uk/q3c/krail/option/api/OptionDao.java
@@ -13,18 +13,16 @@
 
 package uk.q3c.krail.option.api;
 
-import uk.q3c.krail.option.DefaultOptionCacheLoader;
-
 import javax.annotation.Nonnull;
 import java.util.Optional;
 
 /**
  * Provides data access to persistence for {@link OptionDaoDelegate}
  * <br>
- * * <b>NOTE:</b> All values to and from {@link Option} are natively typed.  All values to and from {@link OptionCache}, {@link DefaultOptionCacheLoader} and
+ * * <b>NOTE:</b> All values to and from {@link Option} are natively typed.  All values to and from {@link OptionCache}, DefaultOptionCacheLoader and
  * {@link OptionDaoDelegate} are wrapped in Optional.
  * <p>
- * <p>
+ * <p>Note: DefaultOptionCacheLoader has no interface and is in krail-option
  * Created by David Sowerby on 20/02/15.
  */
 public interface OptionDao {