+++ src/main/java/uk/q3c/krail/option/test/MockOptionPermissionVerifier.java
@@ -1,7 +1,7 @@
 package uk.q3c.krail.option.test;
 
 import org.apache.shiro.authz.AuthorizationException;
-import uk.q3c.krail.core.option.OptionPermission;
+import uk.q3c.krail.option.api.OptionEditAction;
 import uk.q3c.krail.option.api.OptionKey;
 import uk.q3c.krail.option.api.OptionPermissionVerifier;
 import uk.q3c.krail.option.api.UserHierarchy;
@@ -14,7 +14,7 @@ public class MockOptionPermissionVerifier implements OptionPermissionVerifier {
     private boolean throwException = false;
 
     @Override
-    public <T> boolean userHasPermission(OptionPermission.Action action, UserHierarchy hierarchy, int hierarchyRank, OptionKey<T> optionKey) {
+    public <T> boolean userHasPermission(OptionEditAction action, UserHierarchy hierarchy, int hierarchyRank, OptionKey<T> optionKey) {
         if (throwException) {
             throw new AuthorizationException("fake exception");
         }