+++ Release Notes.md
@@ -39,7 +39,9 @@ The `MessageFormat` utility class has been deprecated and replaced by interface 
 - `OptionContext.getOption` is now `OptionContext.optionInstance()`
 - `OptionContext` is parameterised
 - `OptionList` deleted
-- Scope of `DefaultOptionCache` is set in `OptionModule` instead of annotation
+- Scope of `DefaultOptionCache` is set in `OptionModule` instead of by annotation
+- `OptionSource.getContainer()` moved to a new interface `OptionContainerSource` in 'core', to remove Vaadin dependency
+- `OptionPermission.Action` is now `OptionEditAction`
 
 ### Util
 All of `uk.q3c.util` has been reviewed and either allocated to sub-packages or moved to somewhere more appropriate. Those remaining in the util package have also been moved to a separate library, **q3c-util**